+++ tests/test_components.py
@@ -16,7 +16,7 @@ import climt
 from sympl import (
     Stepper, TendencyStepper, TimeDifferencingWrapper,
     ImplicitTendencyComponent, UpdateFrequencyWrapper, DataArray,
-    TendencyComponent, AdamsBashforth, get_constant
+    TendencyComponent, AdamsBashforth
 )
 from sympl._core.tracers import reset_tracers, reset_packers
 from datetime import datetime, timedelta
@@ -445,44 +445,6 @@ class TestInstellation(ComponentBaseColumn, ComponentBase3D):
 class TestDryConvection(ComponentBaseColumn, ComponentBase3D):
     def get_component_instance(self):
         return DryConvectiveAdjustment()
-
-
-def heat_capacity(q):
-
-    Cpd = get_constant('heat_capacity_of_dry_air_at_constant_pressure', 'J/kg/degK')
-    Cvap = get_constant('heat_capacity_of_vapor_phase', 'J/kg/K')
-
-    return Cpd*(1-q) + Cvap*q
-
-
-def test_DryConvectiveAdjustment_conserves_enthalpy_and_water():
-    """
-    Test whether the enthalpy and amount of water is the same after
-    DryConvectiveAdjustment eliminates the instability in the column.
-    """
-
-    unstable_level = 1
-    conv_adj = climt.DryConvectiveAdjustment()
-
-    state = climt.get_default_state([conv_adj], grid_state=climt.get_grid(nz=35))
-
-    dp = (state['air_pressure_on_interface_levels'][:-1] - state['air_pressure_on_interface_levels'][1:])/(
-        state['air_pressure_on_interface_levels'][0] - state['air_pressure_on_interface_levels'][-1])
-    state['air_temperature'][:unstable_level] += 10
-    state['specific_humidity'][:unstable_level] = 0.05
-
-    initial_enthalpy = np.sum(heat_capacity(state['specific_humidity']).values *
-                              state['air_temperature'].values*dp.values)
-    initial_water = np.sum(state['specific_humidity'].values*dp.values)
-
-    diag, output = conv_adj(state, timedelta(hours=1))
-
-    final_water = np.sum(output['specific_humidity'].values*dp.values)
-    final_enthalpy = np.sum(heat_capacity(output['specific_humidity']).values *
-                            output['air_temperature'].values*dp.values)
-
-    assert np.isclose(initial_water, final_water)
-    assert np.isclose(initial_enthalpy, final_enthalpy)
 
 
 @pytest.mark.skip('until get_default_state is fixed')