+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 Latest
 ------
 
+* Add tests to verify conservation of quantities
 * Fix bugs in initialisation
 * Fix energy conservation in surface flux scheme
 * Enable the McICA version of RRTMG Shortwave,